backend/pyproject.toml
[project]
name = "perplex-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "Brennan Beal",email = "brennan@withprotege.ai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.97.1,<2.0.0)",
    "openai-agents (>=0.2.3,<0.3.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "fastapi (>=0.104.1,<1.0.0)",
    "uvicorn[standard] (>=0.24.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic (>=2.5.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
